--- tools/serverpod_cli/lib/src/analyzer/protocol_definition.dart
@@ -1,64 +1,0 @@
-import 'package:analyzer/dart/element/element.dart';
-import 'package:path/path.dart' as p;
-
-import '../util/extensions.dart';
-import '../generator/types.dart';
-
-class ProtocolDefinition {
-  final List<EndpointDefinition> endpoints;
-  final List<String> filePaths;
-
-  ProtocolDefinition({
-    required this.endpoints,
-    required this.filePaths,
-  });
-}
-
-class EndpointDefinition {
-  final String name;
-  final String? documentationComment;
-  final String className;
-  final String fileName;
-  final List<MethodDefinition> methods;
-  final String? subDir;
-
-  EndpointDefinition({
-    required this.name,
-    required this.documentationComment,
-    required this.methods,
-    required this.className,
-    required this.fileName,
-    required this.subDir,
-  });
-}
-
-class MethodDefinition {
-  final String name;
-  final String? documentationComment;
-  final TypeDefinition returnType;
-  final List<ParameterDefinition> parameters;
-  final List<ParameterDefinition> parametersPositional;
-  final List<ParameterDefinition> parametersNamed;
-
-  MethodDefinition(
-      {required this.name,
-      required this.documentationComment,
-      required this.parameters,
-      required this.parametersPositional,
-      required this.parametersNamed,
-      required this.returnType});
-}
-
-class ParameterDefinition {
-  final String name;
-  final TypeDefinition type;
-  final bool required;
-  final ParameterElement? dartParameter;
-
-  ParameterDefinition({
-    required this.name,
-    required this.type,
-    required this.required,
-    this.dartParameter,
-  });
-}
